+++ packages/gov-ui-kit/src/modules/assets/copy/modulesCopy.ts
@@ -33,6 +33,18 @@ export const modulesCopy = {
     proposalActionsAction: {
         notVerified: 'Not verified',
     },
+    proposalActionChangeMembers: {
+        summary: 'Summary',
+        added: 'Added',
+        removed: 'Removed',
+        members: 'members',
+        adjustMemberCount: {
+            addMembers: 'Add members',
+            removeMembers: 'Remove members',
+        },
+        existingMembers: 'Existing members',
+        blockNote: 'At the block in which the proposal was created',
+    },
     proposalActionsUpdateMetadata: {
         logoTerm: 'Logo',
         nameTerm: 'Name',