+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/actions/index.ts
@@ -1,3 +1,4 @@
+export * from './proposalActionChangeMembers';
 export * from './proposalActionTokenMint';
 export * from './proposalActionUpdateMetadata';
 export * from './proposalActionWithdrawToken';