+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/actions/proposalActionChangeMembers/index.ts
@@ -1,0 +1,1 @@
+export { ProposalActionChangeMembers, type IProposalActionChangeMembersProps } from './proposalActionChangeMembers';