+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/proposalActionsUtils.test.tsx
@@ -1,13 +1,19 @@
 import { render, screen } from '@testing-library/react';
-import { generateProposalActionTokenMint, generateProposalActionUpdateMetadata } from './actions/generators';
-import { generateProposalAction } from './actions/generators/proposalAction';
-import { generateProposalActionWithdrawToken } from './actions/generators/proposalActionWithdrawToken';
+import {
+    generateProposalAction,
+    generateProposalActionChangeMembers,
+    generateProposalActionTokenMint,
+    generateProposalActionUpdateMetadata,
+    generateProposalActionWithdrawToken,
+} from './actions/generators';
+import { ProposalActionType } from './proposalActionsTypes';
 import { proposalActionsUtils } from './proposalActionsUtils';
 
 jest.mock('./actions', () => ({
     ProposalActionWithdrawToken: () => <div>Mock ProposalActionWithdrawToken</div>,
     ProposalActionUpdateMetadata: () => <div>Mock ProposalActionUpdateMetaData</div>,
     ProposalActionTokenMint: () => <div>Mock ProposalActionTokenMint</div>,
+    ProposalActionChangeMembers: () => <div>Mock ProposalActionChangeMembers</div>,
 }));
 
 describe('ProposalActions utils', () => {
@@ -33,6 +39,22 @@ describe('ProposalActions utils', () => {
         const Component = proposalActionsUtils.getActionComponent(action)!;
         render(<Component />);
         expect(screen.getByText('Mock ProposalActionTokenMint')).toBeInTheDocument();
+    });
+
+    it('returns ProposalActionChangeMembers component for addMember action', () => {
+        const action = generateProposalActionChangeMembers({ type: ProposalActionType.ADD_MEMBERS });
+
+        const Component = proposalActionsUtils.getActionComponent(action)!;
+        render(<Component />);
+        expect(screen.getByText('Mock ProposalActionChangeMembers')).toBeInTheDocument();
+    });
+
+    it('returns ProposalActionChangeMembers component for removeMember action', () => {
+        const action = generateProposalActionChangeMembers({ type: ProposalActionType.REMOVE_MEMBERS });
+
+        const Component = proposalActionsUtils.getActionComponent(action)!;
+        render(<Component />);
+        expect(screen.getByText('Mock ProposalActionChangeMembers')).toBeInTheDocument();
     });
 
     it('returns null for unknown action type', () => {